Removed dependency on dateutil.parser to convert a UTC timestamp to localtime - Removed unused dirs - Migration: fixed column width for column listing/editing screen - copytable: fixed warning [TESTS] - test function individually in shell - view log files in a 5.7 server, verify timestamps are local

+++ plugins/migration/copytable/copytable.cpp
@@ -269,7 +269,7 @@ RowBuffer::RowBuffer(boost::shared_ptr<std::vector<ColumnInfo> > columns,
           bind.is_null = NULL;
         }
         throw std::runtime_error(base::strfmt("Could not allocate %i bytes for row buffer column of %s %s %i", 
-          bind.buffer_length, col->source_name.c_str(), col->source_type.c_str(), col->target_type));
+          (int)bind.buffer_length, col->source_name.c_str(), col->source_type.c_str(), col->target_type));
       }
     }
     else
